fix: business logic

+++ src/main/java/com/example/restservice/TransactionStatements/dto/CreateTransactionStatementsRequestDTO.java
@@ -1,5 +1,6 @@
 package com.example.restservice.TransactionStatements.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 import com.example.restservice.TransactionStatements.domain.TransactionStatementsMethod;
@@ -15,7 +16,7 @@ public record CreateTransactionStatementsRequestDTO(
     @NotNull(message = "User ID is required") UUID userId,
     UUID orderId,
     @NotNull(message = "Amount is required") @Positive(message = "Amount must be greater than zero")
-        Double amount,
+        BigDecimal amount,
     @NotNull(message = "Transaction type is required") TransactionStatementsType type,
     @NotNull(message = "Transaction method is required") TransactionStatementsMethod method,
     @NotNull(message = "Transaction status is required") TransactionStatementsStatus status,
